Extract grouped aggregate state
